WorkshopAgent/requirements.txt
google-adk==1.19.0
google-cloud-aiplatform==1.128.0
google-genai==1.52.0
google-auth
google-auth-oauthlib
requests
nest-asyncio